requirements.txt
ccxt
python-telegram-bot
python-dotenv
pandas
ta
SQLAlchemy
numpy